+++ src/app/components/SpaceDetail.svelte
@@ -34,21 +34,29 @@
 </script>
 
 <div class="column gap-4">
-  <div class="relative flex gap-4">
-    <div class="relative">
-      <div class="avatar relative">
-        <div
-          class="center !flex h-16 w-16 min-w-16 rounded-full border-2 border-solid border-base-300 bg-base-300">
-          <RelayIcon {url} size={10} />
+  <div class="flex justify-between">
+    <div class="relative flex gap-4">
+      <div class="relative">
+        <div class="avatar relative">
+          <div
+            class="center !flex h-16 w-16 min-w-16 rounded-full border-2 border-solid border-base-300 bg-base-300">
+            <RelayIcon {url} size={10} />
+          </div>
         </div>
       </div>
+      <div class="flex min-w-0 flex-col gap-1">
+        <h1 class="ellipsize whitespace-nowrap text-2xl font-bold">
+          <RelayName {url} />
+        </h1>
+        <p class="ellipsize text-sm opacity-75">{displayRelayUrl(url)}</p>
+      </div>
     </div>
-    <div class="flex min-w-0 flex-col gap-1">
-      <h1 class="ellipsize whitespace-nowrap text-2xl font-bold">
-        <RelayName {url} />
-      </h1>
-      <p class="ellipsize text-sm opacity-75">{displayRelayUrl(url)}</p>
-    </div>
+    {#if $userIsAdmin}
+      <Button class="btn btn-primary" onclick={startEdit}>
+        <Icon icon={Pen} />
+        Edit
+      </Button>
+    {/if}
   </div>
   <RelayDescription {url} />
   {#if $relay?.terms_of_service || $relay?.privacy_policy}
@@ -83,18 +91,10 @@
       </div>
     {/if}
   </div>
-  {#if $userIsAdmin}
-    <ModalFooter>
-      <Button class="btn btn-link" onclick={back}>
-        <Icon icon={AltArrowLeft} />
-        Go back
-      </Button>
-      <Button class="btn btn-primary" onclick={startEdit}>
-        <Icon icon={Pen} />
-        Edit Space
-      </Button>
-    </ModalFooter>
-  {:else}
-    <Button class="btn btn-primary" onclick={back}>Got it</Button>
-  {/if}
+  <ModalFooter>
+    <Button class="btn btn-link" onclick={back}>
+      <Icon icon={AltArrowLeft} />
+      Go back
+    </Button>
+  </ModalFooter>
 </div>